requirements.txt
# Part 1: Data Acquisition
requests>=2.31.0
overpy>=0.7
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
shapely>=2.0.0
geopandas>=0.14.0
rasterio>=1.3.0
opencv-python>=4.8.0
pyproj>=3.6.0

# Part 2: Machine Learning
# NOTE: TensorFlow removed - not used in codebase
torch>=2.0.0,<2.3.0
torchvision>=0.15.0,<0.18.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pandas>=2.0.0
albumentations>=1.3.0
tensorboard>=2.13.0

# Part 3: Executable & OSM Integration
osmapi>=4.0.0
PyQt5>=5.15.0
pyinstaller>=6.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Testing (Development)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Security
safety>=2.3.5  # Dependency vulnerability scanning
